+++ include/aquarius/core/generator.hpp
@@ -1,0 +1,31 @@
+#pragma once
+#include <aquarius/core/impl/router.hpp>
+#include <aquarius/core/singleton.hpp>
+#include <functional>
+
+namespace aquarius
+{
+	template <typename _R, typename... _Args>
+	class generator : public impl::router<std::size_t, std::function<_R(_Args...)>>
+	{
+	public:
+		generator() = default;
+
+	public:
+		_R invoke(std::size_t key, _Args&... args)
+		{
+			std::lock_guard lk(this->mutex_);
+
+			auto iter = this->map_invokes_.find(key);
+
+			if (iter == this->map_invokes_.end())
+				return _R();
+
+			return iter->second(args...);
+		}
+	};
+
+	template <typename _Single, typename _R, typename... _Args>
+	class single_generator : public generator<_R, _Args...>, public singleton<_Single>
+	{};
+} // namespace aquarius
