+++ include/aquarius/service.hpp
@@ -1,21 +1,19 @@
 #pragma once
 #include <string>
+#include <aquarius/service/service.hpp>
+#include <aquarius/service/manager.hpp>
 
 namespace aquarius
 {
-	class service
+	template <typename _Service>
+	struct service_regist
 	{
-	public:
-		virtual bool init() = 0;
-
-		virtual bool config() = 0;
-
-		virtual bool run() = 0;
-
-		virtual void stop() = 0;
-
-		virtual bool enable() = 0;
-
-		virtual std::string name() = 0;
+		service_regist(std::size_t service_name)
+		{
+			service_manager::instance().regist(service_name, std::make_shared<_Service>());
+		}
 	};
+
 } // namespace aquarius
+
+#define AQUARIUS_SERVICE_REGIST(name, service) static aquarius::service_regist<service> srv_##service(name)
