+++ test/unit_test.vcxproj
@@ -162,42 +162,43 @@
     <ClInclude Include="..\include\aquarius\channel\publisher.hpp" />
     <ClInclude Include="..\include\aquarius\channel\subscriber.hpp" />
     <ClInclude Include="..\include\aquarius\client.hpp" />
+    <ClInclude Include="..\include\aquarius\client\async.hpp" />
+    <ClInclude Include="..\include\aquarius\client\callback.hpp" />
     <ClInclude Include="..\include\aquarius\connect.hpp" />
     <ClInclude Include="..\include\aquarius\connect\impl\session.hpp" />
+    <ClInclude Include="..\include\aquarius\connect\invoke.hpp" />
+    <ClInclude Include="..\include\aquarius\connect\manager.hpp" />
     <ClInclude Include="..\include\aquarius\connect\session.hpp" />
     <ClInclude Include="..\include\aquarius\connect\impl\socket_adapter.hpp" />
     <ClInclude Include="..\include\aquarius\context.hpp" />
+    <ClInclude Include="..\include\aquarius\context\content.hpp" />
     <ClInclude Include="..\include\aquarius\context\context.hpp" />
+    <ClInclude Include="..\include\aquarius\context\generator.hpp" />
+    <ClInclude Include="..\include\aquarius\context\impl\context.hpp" />
     <ClInclude Include="..\include\aquarius\context\impl\visitor.hpp" />
+    <ClInclude Include="..\include\aquarius\context\invoke.hpp" />
+    <ClInclude Include="..\include\aquarius\core\apply.hpp" />
     <ClInclude Include="..\include\aquarius\core\error_code.hpp" />
+    <ClInclude Include="..\include\aquarius\core\generator.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\error_category.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\router.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\system_errc.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\system_error_message.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\time_traits.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\type_traits.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\callback.hpp" />
+    <ClInclude Include="..\include\aquarius\core\manager.hpp" />
+    <ClInclude Include="..\include\aquarius\message\generator.hpp" />
     <ClInclude Include="..\include\aquarius\message\header.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\custom.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\header.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\field.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\protocol.hpp" />
-    <ClInclude Include="..\include\aquarius\router\context.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\async.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\apply.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\generator.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\manager.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\router.hpp" />
     <ClInclude Include="..\include\aquarius\core\singleton.hpp" />
-    <ClInclude Include="..\include\aquarius\router\message.hpp" />
-    <ClInclude Include="..\include\aquarius\router\service.hpp" />
-    <ClInclude Include="..\include\aquarius\router\session.hpp" />
+    <ClInclude Include="..\include\aquarius\message\invoke.hpp" />
+    <ClInclude Include="..\include\aquarius\router\router.hpp" />
     <ClInclude Include="..\include\aquarius\server\io_service_pool.hpp" />
     <ClInclude Include="..\include\aquarius\core\core.hpp" />
     <ClInclude Include="..\include\aquarius\core\elastic.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\context.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\message.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\service.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\session.hpp" />
     <ClInclude Include="..\include\aquarius\core\logger.hpp" />
     <ClInclude Include="..\include\aquarius\message\message.hpp" />
     <ClInclude Include="..\include\aquarius\request.hpp" />
@@ -211,6 +212,8 @@
     <ClInclude Include="..\include\aquarius\core\uuid.hpp" />
     <ClInclude Include="..\include\aquarius\core\concepts.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\visitable.hpp" />
+    <ClInclude Include="..\include\aquarius\service\manager.hpp" />
+    <ClInclude Include="..\include\aquarius\service\service.hpp" />
     <ClInclude Include="context.h" />
     <ClInclude Include="error_code.h" />
     <ClInclude Include="invoke.h" />
